Add debugDescription for SyntaxTree
It seems to work fine. It is not well tested, but I think it is enough at the moment. I have an idea to make code generation based on grammar to represent syntax tree in strict and safe form, but at first I need to accomplish my main goal: mock framework.